Add eslint; Move files to clearer places
