+++ src/performance/monitor/service.js
@@ -8,72 +8,72 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-const requestService = require('./request.service');
-const { repeatFunc } = require('./utils.service');
+const { getResData } = require('../utils/request.service');
+const { repeatFunc } = require('../utils/utils.service');
 
 const numPolls = 100;
 const latestProjectData = {};
 
-const getEnvData = (appOrigin) => requestService.getResData(appOrigin, '/metrics/codewind/environment');
+const getEnvData = (appOrigin) => getResData(appOrigin, '/metrics/codewind/environment');
 
 const getMetricsFromProject = async (appOrigin) => {
-    const projectData = latestProjectData[appOrigin];
-    if (!projectData.hasOwnProperty('metrics')) {
-        projectData.metrics = {};
-    }
-    const [
-        metricsFromUser,
-        metricsFromCodewind,
-    ] = await Promise.all([
-        requestService.getResData(appOrigin, '/metrics')
-            .catch(err => console.log(err.message)),
-        requestService.getResData(appOrigin, '/metrics/codewind')
-            .catch(err => console.log(err.message)),
-    ]);
+  const projectData = latestProjectData[appOrigin];
+  if (!projectData.hasOwnProperty('metrics')) {
+    projectData.metrics = {}
+  }
+  const [
+    metricsFromUser,
+    metricsFromCodewind,
+  ] = await Promise.all([
+    getResData(appOrigin, '/metrics')
+      .catch(err => console.log(err.message)),
+    getResData(appOrigin, '/metrics/codewind')
+      .catch(err => console.log(err.message)),
+  ]);
 
-    projectData.metrics.fromUser = metricsFromUser;
-    projectData.metrics.fromCodewind = metricsFromCodewind;
+  projectData.metrics.fromUser = metricsFromUser;
+  projectData.metrics.fromCodewind = metricsFromCodewind;
 }
 
 const getProfilingDataFromProject = async (appOrigin) => {
-    const projectData = latestProjectData[appOrigin];
-    if (!projectData.hasOwnProperty('profiling')) {
-        projectData.profiling = {};
-    }
-    const profilingData = await requestService.getResData(appOrigin, '/metrics/codewind/profiling');
-    projectData.profiling = profilingData;
+  const projectData = latestProjectData[appOrigin];
+  if (!projectData.hasOwnProperty('profiling')) {
+    projectData.profiling = {};
+  }
+  const profilingData = await getResData(appOrigin, '/metrics/codewind/profiling');
+  projectData.profiling = profilingData;
 }
 
-const scrapeProjectData = async(appOrigin, projectLanguage) => {
-    if (!latestProjectData.hasOwnProperty(appOrigin)) {
-        latestProjectData[appOrigin] = {};
-    }
-    const promises = [
-        repeatFunc(
-            () => getMetricsFromProject(appOrigin),
-            numPolls,
-        ),
-    ]
-    if (projectLanguage === 'nodejs') {
-        promises.push(
-            repeatFunc(
-                () => getProfilingDataFromProject(appOrigin),
-                numPolls,
-            )
-        );
-    }
-    await Promise.all(promises);
+const scrapeProjectData = async (appOrigin, projectLanguage) => {
+  if (!latestProjectData.hasOwnProperty(appOrigin)) {
+    latestProjectData[appOrigin] = {};
+  }
+  const promises = [
+    repeatFunc(
+      () => getMetricsFromProject(appOrigin),
+      numPolls,
+    ),
+  ]
+  if (projectLanguage === 'nodejs') {
+    promises.push(
+      repeatFunc(
+        () => getProfilingDataFromProject(appOrigin),
+        numPolls,
+      )
+    );
+  }
+  await Promise.all(promises);
 }
 
 const getLatestProjectData = (appOrigin) => latestProjectData[appOrigin];
 
 const resetProjectData = (appOrigin) => {
-    latestProjectData[appOrigin] = {};
+  latestProjectData[appOrigin] = {};
 }
 
 module.exports = {
-    resetProjectData,
-    getLatestProjectData,
-    scrapeProjectData,
-    getEnvData,
+  resetProjectData,
+  getLatestProjectData,
+  scrapeProjectData,
+  getEnvData,
 };
